feat: Refac, Seperate cli pkg
